adjust date picking also model and controller, confirm when click delete btn

+++ frontend/src/pages/admin/OrdersList.jsx
@@ -44,7 +44,9 @@ const OrdersList = () => {
   }, [dispatch, error, alert, deleteError, isDeleted, navigate]);
 
   const deleteOrderHandler = (id) => {
-    dispatch(deleteOrder(id));
+    if (window.confirm("Có chắc chắn xóa?")) {
+      dispatch(deleteOrder(id))
+    }
   };
 
   const columns = [
